[Design] Fix PlaceDropdown conponent location and Fix Mobile Header
